// src/interfaces/order.ts
export interface OrderInterface {
    Rx_Number: string,
    Order_Number: string;
    Patient_Name: string;
    Prescriber_Name: string;
    Status: string;
    tracking_code: string;
    Tracking: string;
    Medication: string;
    Issue_Date: string;
    fname: string;
    lname: string;
    Order_id: string;
    order_id: string;
    ship_cn: string;
    acct_id: string;
    md_npi: string;
    order_add_date: string;
    birth_date: string;
    md_first_name: string;
    md_last_name: string;
    addr1: string;
    city: string;
    state_cd: string;
    zip: string;
    item_name: string;
    rx_price: string;
    NeworRefill: string;
    written_qty: string;
    sig_text: string;
    script_no: string;
    Invoice_Nbr:string
}